+++ ToDoApplication/Mapping/ShoppingProductMappingProfile.cs
@@ -1,0 +1,15 @@
+﻿using AutoMapper;
+using ToDoApplication.Models;
+using ToDoApplication.ViewModels;
+
+namespace ToDoApplication.MappingProfiles
+{
+    public class ShoppingProductMappingProfile : Profile
+    {
+        public ShoppingProductMappingProfile()
+        {
+            CreateMap<ShoppingProduct, ShoppingProductViewModel>();
+            CreateMap<ShoppingProductViewModel, ShoppingProduct>();
+        }
+    }
+}
