+++ ToDoApplication/Mapping/TaskToDoMappingProfile.cs
@@ -1,0 +1,15 @@
+﻿using AutoMapper;
+using ToDoApplication.Models;
+using ToDoApplication.ViewModels;
+
+namespace ToDoApplication.MappingProfiles
+{
+    public class TaskToDoMappingProfile : Profile
+    {
+        public TaskToDoMappingProfile()
+        {
+            CreateMap<TaskToDo, TaskToDoViewModel>();
+            CreateMap<TaskToDoViewModel, TaskToDo>();
+        }
+    }
+}
