--- ToDoApplication/MappingProfiles/TaskToDoMappingProfile.cs
@@ -1,32 +1,0 @@
-﻿using AutoMapper;
-using ToDoApplication.Models;
-using ToDoApplication.ViewModels;
-
-namespace ToDoApplication.MappingProfiles
-{
-    public class TaskToDoMappingProfile : Profile
-    {
-        public TaskToDoMappingProfile()
-        {
-            CreateMap<TaskToDo, TaskToDoViewModel>()
-                .ForMember(dest => dest.TaskId, opt => opt.MapFrom(src => src.TaskId))
-                .ForMember(dest => dest.Name, opt => opt.MapFrom(src => src.Name))
-                .ForMember(dest => dest.Description, opt => opt.MapFrom(src => src.Description))
-                .ForMember(dest => dest.Date, opt => opt.MapFrom(src => src.Date))
-                .ForMember(dest => dest.Time, opt => opt.MapFrom(src => src.Time))
-                .ForMember(dest => dest.Category, opt => opt.MapFrom(src => src.Category))
-                .ForMember(dest => dest.Status, opt => opt.MapFrom(src => src.Status))
-                .ForMember(dest => dest.LastDone, opt => opt.MapFrom(src => src.LastDone));
-
-            CreateMap<TaskToDoViewModel, TaskToDo>()
-                .ForMember(dest => dest.TaskId, opt => opt.MapFrom(src => src.TaskId))
-                .ForMember(dest => dest.Name, opt => opt.MapFrom(src => src.Name))
-                .ForMember(dest => dest.Description, opt => opt.MapFrom(src => src.Description))
-                .ForMember(dest => dest.Date, opt => opt.MapFrom(src => src.Date))
-                .ForMember(dest => dest.Time, opt => opt.MapFrom(src => src.Time))
-                .ForMember(dest => dest.Category, opt => opt.MapFrom(src => src.Category))
-                .ForMember(dest => dest.Status, opt => opt.MapFrom(src => src.Status))
-                .ForMember(dest => dest.LastDone, opt => opt.MapFrom(src => src.LastDone));
-        }
-    }
-}
